# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RentX Documentation",
        "description": "This is an API Rent",
        "version": "1.0.0",
        "contact": {
            "email": "renato.hossotani@gmail.com"
        }
    },
    "paths": {
        "/categories": {
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Create a category",
                "description": "Create a new category",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                }
                            },
                            "example": {
                                "name": "category name sample",
                                "description": "category description sample"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "500": {
                        "description": "Category already exists"
                    }
                }
            },
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "List all categories",
                "description": "List all categories",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "aplication/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/categories/import": {
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Upload a new category",
                "description": "Upload a neew category",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/specifications": {
            "post": {
                "tags": [
                    "Specifications"
                ],
                "summary": "Create a specification",
                "description": "Create a new specification",
                "requestBody": {
                    "content": {
                        "aplication/json": {
                            "schema": {
                                "$ref": "#/definitions/Specification"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "500": {
                        "description": "Specification already exists"
                    }
                }
            }
        }
    },
    "definitions": {
        "Specification": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                }
            }
        }
    }
}